==> 1-stage/1_demand-decode-dicofre.ipynb <==
In this notebook we'll obtain the **destination parish from the DICOFRE** codes. The output will be a demand table with the a column for destination parish, instead of DICOFRE code.

import pandas as pd

Churn_OD = pd.read_csv('https://wdl-data.fra1.digitaloceanspaces.com/pse/Churn_OD.txt',
                                 encoding = "ISO-8859-1",
                                sep='|', dtype={'Dicofre_ParishCode_of_Public_Transportation': str})

Obtaining information to decode the DICOFRE code from Portugal's public open access data

parishes = pd.read_excel('https://dados.gov.pt/pt/datasets/r/ec6ef805-c278-4b4d-ba9b-3116264f68b4', engine="openpyxl")[['dicofre', 'freguesia']]

Merging the data

od_parish = pd.merge(Churn_OD, parishes,
                    how='outer',left_on=['Dicofre_ParishCode_of_Public_Transportation'],
                    right_on=['dicofre']).dropna().drop(['dicofre', 'Dicofre_ParishCode_of_Public_Transportation'],1)

Renaming parish column and moving Demand to leftmost position

od_parish=od_parish.rename(columns={'freguesia':'Parish_of_Public_Transportation'})

new_cols = [col for col in od_parish.columns if col != 'Demand_weight'] + ['Demand_weight']
od_parish = od_parish[new_cols]

Saving to csv

od_parish.to_csv('intermediate-data/demand.csv', index=False)
