==> 1-stage/4_variation-exploration.ipynb <==
import pandas as pd
import fiona
import matplotlib.pyplot as plt
pd.options.display.max_rows = 4000
from IPython.display import display
from matplotlib.collections import PatchCollection
from descartes import PolygonPatch
from shapely.geometry import Polygon, MultiPolygon, shape

variation = pd.read_csv('intermediate-data/variation.csv')
demand = pd.read_csv('/home/primity/wdl/1-stage/intermediate-data/demand.csv')

variation.County_of_Origin=variation.County_of_Origin.str.upper()
variation.District_of_Origin=variation.District_of_Origin.str.upper()

demand.County_of_Origin=demand.County_of_Origin.str.upper()
demand.District_of_Origin=demand.District_of_Origin.str.upper()

Demand_weight is relative to the county of Origin

variation.sample(2)

# Groupby of Absolute and Relative Changes for some features
for col in ['Region_of_Origin','District_of_Origin', 'County_of_Origin', 'GenderDescription', 'AgeClassDescription']:
    first_period = variation.groupby(col)['Average_BusUsers_per_Day_first'].sum().to_frame().reset_index().sort_values(col, ascending=False)
    change = variation.groupby(col)['Variation_abs'].sum().to_frame().reset_index().sort_values(col, ascending=False)
    change['Variation_rel'] = change['Variation_abs'] / first_period['Average_BusUsers_per_Day_first']
    display(change.sort_values('Variation_rel', ascending=False))

variation.groupby(['County_of_Origin', ])['Average_BusUsers_per_Day_first'].sum().to_frame().reset_index().sort_values(col, ascending=False)

# Groupby of Absolute and Relative Changes for some features
for col in ['Region_of_Origin','District_of_Origin', 'County_of_Origin', 'GenderDescription', 'AgeClassDescription']:
    first_period = variation.groupby(col)['Average_BusUsers_per_Day_first'].sum().to_frame().reset_index().sort_values(col, ascending=False)
    change = variation.groupby(col)['Variation_abs'].sum().to_frame().reset_index().sort_values(col, ascending=False)
    change['Variation_rel'] = change['Variation_abs'] / first_period['Average_BusUsers_per_Day_first']
    display(change.sort_values('Variation_rel', ascending=False))

In absolute terms, churn was higher in the Porto region, with the district of Aveiro presenting the lowest churn followed by Setúbal and Lisboa, with the Porto district showing the biggest churn. Men and women churned in similar quantities, churn was largest for people of older age, with an exception for the age band of 35-44, wich churned the least.  
For relative churn, which is likely more important, the Porto region also saw bigger churn than Lisbon. As for districts, Aveiro saw relative churns very similar to Porto, which had the largest churn. Males churned around 4% more than females, and for age range the highest-churning were people aged 25-34.

variation.sort_values('Variation_rel').head(3)

# Looping counties of origin with worst churn
for i, county in enumerate(list(variation.sort_values('Variation_rel')['County_of_Origin'])):
    # For those counties, is there demand for them as destination?
    if len(demand[demand.County_of_Public_Transportation==county]):
        print(i, county)
        display(demand[demand.County_of_Public_Transportation==county].sort_values('Demand_weight', ascending=False).head(2))

origins= set(demand.District_of_Origin)
destinations = set(demand.District_of_Public_Transportation)
print('\033[1mDistricts in Origins but not in Destination \033[0m- ', end='')
print(origins-destinations, end= ' ')
print(' ')
print('\033[1mDistricts in Destination but not in Origin \033[0m- ', end='')
print(destinations-origins, end= ' ')
print('\n')
origins= set(demand.County_of_Origin)
destinations = set(demand.County_of_Public_Transportation)
print('\033[1mCounties in Origins but not in Destination \033[0m- ', end='')
print(origins-destinations, end= ' ')
print(' ')
print('\033[1mCounties in Destination but not in Origin \033[0m- ', end='')
print(destinations-origins, end= ' ')

# Parishes that dominate their county

demand.sample(4)

demand.sort_values('Demand_weight', ascending=False)['Demand_weight']
